fix wallpaper setting
Previously setbg used to pass original wallpaper to wal which lead
to it storing the full path to the original file rather than a local
copy. That parameter was used in sway for example, meaning that if
you delete the original file the wallpaper will stop showing up.
So fix that by passing the copied file as a parameter